move TaskParam back next to BaseHandler
