+++ src/handlers_psutil_processes.py
@@ -14,7 +14,7 @@ from typing import (
 
 from .handlers_base import BaseHandler, Payload
 from .utils import string_from_dict_optionally, string_from_list_optionally
-from .task import TaskParam
+from .handlers_base import TaskParam
 
 class ProcessesCommandHandler(BaseHandler):
     '''
