+++ src/handlers_pysmart.py
@@ -8,7 +8,7 @@ from typing import (
 
 from .handlers_base import BaseHandler, Payload
 from .utils import string_from_dict_optionally
-from .task import TaskParam
+from .handlers_base import TaskParam
 
 class SmartCommandHandler(BaseHandler):
     '''
